pyproject.toml
[tool.poetry]
name = "personal-budget"
version = "1"
description = "parses transactions"
authors = ["Ben Caldwell <ben.caldwell@messagemedia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
gspread = "^6.0.2"
oauth2client = "^4.1.3"
thefuzz = "^0.22.1"
click = "^8.1.7"
rich-click = "^1.7.4"
python-dateutil = "^2.9.0.post0"
toml = "^0.10.2"
dependency-injector-fork = "^4.42.1"
structlog = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
